requirements.txt
peft==0.13.2
transformers==4.46.3
huggingface-hub==0.28.1   
accelerate==1.0.1
tokenizers==0.20.3
numpy==1.24.4
polars==1.8.2
scikit-learn==1.3.2
datasets==3.1.0           
wandb==0.19.6
tqdm==4.67.1
matplotlib==3.7.5
seaborn==0.13.2
biopython==1.78
mpi4py==4.0.2
scipy==1.10.1
ete3==3.1.1
evaluate==0.4.6
lightgbm==4.6.0
cairosvg==2.7.1
joblib==1.4.2
six==1.17.0
ipykernel==6.29.5
ipywidgets==8.1.8